4.omo-superpowers/python/omo/backend/pyproject.toml
[project]
name = "cms-backend"
version = "0.1.0"
description = "CMS Backend API"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.5.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "markdown>=3.5.0",
    "bleach>=6.1.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "httpx>=0.26.0",
]